generator: Look at the previously written Components file to decide if we need to update
It's not enough to see if there are new components, since things can
move between suites, for example when transitioning from unstable to
testing. Instead we now read the previous Components file, if any, and
write a new one if we either find a package that wasn't in there before,
or don't find a package that was (deletions).